+++ common/src/main/java/top/vmctcn/vmtranslationupdate/util/VersionCheckUtil.java
@@ -1,8 +1,5 @@
 package top.vmctcn.vmtranslationupdate.util;
 
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.text.Text;
-
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.net.URL;
@@ -10,7 +7,7 @@ import java.net.URLConnection;
 import java.nio.charset.StandardCharsets;
 
 public class VersionCheckUtil {
-    public static String getOnlineVersion(PlayerEntity player) {
+    public static String getOnlineVersion() {
         try {
             URL url = new URL(ModConfigUtil.getConfig().modPackTranslationUpdateCheckUrl);
             URLConnection connection = url.openConnection();
@@ -23,7 +20,6 @@ public class VersionCheckUtil {
                 return reader.readLine();
             }
         } catch (Exception e) {
-            player.sendMessage(Text.translatable("vmtranslationupdate.message.error"));
             return "";
         }
     }