+++ core/src/main/java/top/vmctcn/vmtucore/ModPlatform.java
@@ -1,0 +1,9 @@
+package top.vmctcn.vmtucore;
+
+import top.vmctcn.vmtucore.util.ServiceHelper;
+
+public interface ModPlatform {
+    ModPlatform INSTANCE = ServiceHelper.loadService(ModPlatform.class);
+
+    String getGameVersion();
+}