+++ core/src/main/java/top/vmctcn/vmtucore/respack/metadata/GameMetadata.java
@@ -1,0 +1,14 @@
+package top.vmctcn.vmtucore.respack.metadata;
+
+import java.util.List;
+
+/**
+ * Under AGPL License
+ *
+ * @author CFPAOrg/I18nUpdateMod3 (xfl03)
+ */
+public class GameMetadata {
+    public String gameVersions;
+    public int packFormat;
+    public List<String> convertFrom;
+}